refactor to combine webview and widget window
This commit includes work to make it easier to work with the global
window.

This commit also addresses an issue where functions sent on the webview
sender channel were not processed immediately, but only after a future
window message (generated by the OS) was received. This ended up causing
the function to be called at inconsistent times. The fix is that we now
send the WM_APP message (an application defined message) to the window
immediately after sending on the channel.